🧑🏾‍🔧 Added current events reducer and extracted data to file
