+++ src/components/Events/style.css
@@ -7,7 +7,7 @@
   padding: 2rem 0;
   height: 40rem;
   gap: 2rem;
-  height: 100%;
+  height: 50vh;
 }
 
 .text-white {
